Act on a failed project deletion
Set any FailedPrecondition (of which project-with-rules is one)
to display an error in the delete modal,
while all other errors will close the modal and display
the error in the standard banner.

While working through this I discovered that the gRPC
FailedPrecondition does NOT map to the HTTP FailedPrecondition.
I have noted that in the types.ts file.

Signed-off-by: michael sorens <msorens@chef.io>