[gateway] bind unknown gRPC to an unimplemented backend
As we move to a world where Chef Automate becomes a platform and the
products deployed on any given instance will vary wildly, we need the
gateway to be able to initialize with several different combinations of
binds, depending on the services that are required by a given product.

Previously, the gateway had a hardcoded list of target addresses and
ports which it would use to dial in cases where target information for a
service was not present in the configuration. If we were to start up the
gateway with only the binds required by a small product it would still
attempt to dial any gRPC client that it might need, regardless of if the
client was ever going to exist. This leads to many failures and noisey
logs.

This change introduces a few new concepts:

* Instead of always trying to dial all services from a known "best
  guess" list, we instead dial configured targets and for each service
  that is missing in the configuration, we'll dial the "null backend".
  The null backend is an aggregate Chef Automate backend that implements
  all client RPCS and returns "unimplemented" errors. This will allow
  the gateway to cleanly initialize regardless of the given binds and
  will return clean errors if an RPC is requested and the required
  backend is not available.

* The gateway now handles signals and attempts to cleanly shut down when
  terminated. It also now responds to SIGHUP and will reload the
  configuration and apply it without restarting.

This change also:

* Add optional binds for all services which the gateway migth attempt to
  dial.

* Replace functional options for the NewServer and replace them with the
  Config. In most cases it's much less work to reference the Config than
  use functional options to set equivalent fields on the Server struct.

* Decomposes `Start` into several smaller functions, allowing better
  control and re-usability when reconfiguring.

Signed-off-by: Ryan Cragun <ryan@chef.io>